remove stutter in name
