+++ notification/messagetype.go
@@ -19,7 +19,7 @@ const (
 	MessageTypeVerification
 	MessageTypeAlertBundle
 	MessageTypeAlertStatusBundle
-	MessageTypeScheduleOnCallNotification
+	MessageTypeScheduleOnCallUsers
 )
 
 func (s MessageType) Value() (driver.Value, error) {
@@ -36,7 +36,7 @@ func (s MessageType) Value() (driver.Value, error) {
 		return "alert_notification_bundle", nil
 	case MessageTypeAlertStatusBundle:
 		return "alert_status_update_bundle", nil
-	case MessageTypeScheduleOnCallNotification:
+	case MessageTypeScheduleOnCallUsers:
 		return "schedule_on_call_notification", nil
 	}
 	return nil, fmt.Errorf("could not process unknown type for MessageType %s", s)
@@ -59,7 +59,7 @@ func (s *MessageType) Scan(value interface{}) error {
 	case "alert_status_update_bundle":
 		*s = MessageTypeAlertStatusBundle
 	case "schedule_on_call_notification":
-		*s = MessageTypeScheduleOnCallNotification
+		*s = MessageTypeScheduleOnCallUsers
 	default:
 		return fmt.Errorf("could not process unknown type for MessageType %str", str)
 	}
