+++ notification/slack/channel.go
@@ -322,7 +322,7 @@ func (s *ChannelSender) Send(ctx context.Context, msg notification.Message) (str
 		vals.Set("text", fmt.Sprintf("Alert: %s\n\n<%s>", t.Summary, cfg.CallbackURL("/alerts/"+strconv.Itoa(t.AlertID))))
 	case notification.AlertBundle:
 		vals.Set("text", fmt.Sprintf("Service '%s' has %d unacknowledged alerts.\n\n<%s>", t.ServiceName, t.Count, cfg.CallbackURL("/services/"+t.ServiceID+"/alerts")))
-	case notification.ScheduleOnCallNotification:
+	case notification.ScheduleOnCallUsers:
 		var userStr string
 		if len(t.Users) == 0 {
 			userStr = "No users"
